stakepoold: Move rpc/rpcserver to rpc/server (#550)
