+++ backend/stakepoold/rpc/server/log.go
@@ -3,7 +3,7 @@
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
-package rpcserver
+package server
 
 import "github.com/decred/slog"
